Tons of changes towards getting code to compile
