+++ include/mapbox/geometry/wagyu/point.hpp
@@ -35,7 +35,7 @@ struct point {
 
     point() : ring(nullptr), x(0), y(0), prev(this), next(this) {
     }
-    point(T x_, T y_) : index(0), x(x_), y(y_), next(this), prev(this) {
+    point(T x_, T y_) : ring(nullptr), x(x_), y(y_), next(this), prev(this) {
     }
     point(ring_ptr<T> ring_, mapbox::geometry::point<T> const& pt)
         : ring(ring_), x(pt.x), y(pt.y), next(this), prev(this) {
