+++ include/mapbox/geometry/wagyu/process_horizontal.hpp
@@ -15,9 +15,26 @@ namespace geometry {
 namespace wagyu {
 
 template <typename T>
-point_ptr<T> get_last_point(edge_ptr<T> e, ring_list<T> const& rings) {
-    ring_ptr<T> outRec = rings[e->index];
-    if (e->side == edge_left) {
+bool horizontal_segments_overlap(T seg1a, T seg1b, T seg2a, T seg2b) {
+    if (seg1a > seg1b) {
+        if (seg2a > seg2b) {
+            return (seg1b < seg2a) && (seg2b < seg1a);
+        } else {
+            return (seg1b < seg2b) && (seg2a < seg1a);
+        }
+    } else {
+        if (seg2a > seg2b) {
+            return (seg1a < seg2a) && (seg2b < seg1b);
+        } else {
+            return (seg1a < seg2b) && (seg2a < seg1b);
+        }
+    }
+}
+
+template <typename T>
+point_ptr<T> get_last_point(active_bound_list_itr<T>& bnd) {
+    ring_ptr<T> outRec = (*bnd)->ring;
+    if ((*bnd)->side == edge_left) {
         return outRec->points;
     } else {
         return outRec->points->prev;
@@ -25,235 +42,365 @@ point_ptr<T> get_last_point(edge_ptr<T> e, ring_list<T> const& rings) {
 }
 
 template <typename T>
-void get_horizontal_direction(edge_ptr<T> edge, horizontal_direction& dir, T& left, T& right) {
-    if (edge->bot.x < edge->top.x) {
-        left = edge->bot.x;
-        right = edge->top.x;
-        dir = left_to_right;
+void process_horizontal_left_to_right(T scanline_y,
+                            active_bound_list_itr<T> horz_bound,
+                            maxima_list<T>& maxima,
+                            active_bound_list<T>& active_bounds,
+                            join_list<T>& joins,
+                            ring_list<T>& rings,
+                            scanbeam_list<T>& scanbeam,
+                            clip_type cliptype,
+                            fill_type subject_fill_type,
+                            fill_type clip_fill_type) {
+    bool is_open = (*horz_bound)->winding_delta == 0;
+    bool is_maxima_edge = is_maxima(horz_bound, scanline_y);
+    auto bound_max_pair = active_bounds.end();
+    if (is_maxima_edge) {
+        bound_max_pair = get_maxima_pair<T>(horz_bound, active_bounds);
+    }
+
+    auto max_iter = maxima.begin();
+    while (max_iter != maxima.end() && *max_iter <= (*horz_bound)->current_edge->bot.x) {
+        ++max_iter;
+    }
+
+    point_ptr<T> p1 = nullptr;
+
+    auto bnd = std::next(horz_bound);
+
+    while (bnd != active_bounds.end()) {
+        // this code block inserts extra coords into horizontal edges (in output
+        // polygons) wherever maxima touch these horizontal edges. This helps
+        //'simplifying' polygons (ie if the Simplify property is set).
+        while (max_iter != maxima.end() && *max_iter < (*bnd)->curr.x) {
+            if ((*horz_bound)->ring && !is_open) {
+                add_point_to_ring(horz_bound,
+                          mapbox::geometry::point<T>(*max_iter, (*horz_bound)->current_edge->bot.y));
+            }
+            ++max_iter;
+        }
+
+        if ((*bnd)->curr.x > (*horz_bound)->current_edge->top.x) {
+            break;
+        }
+
+        // Also break if we've got to the end of an intermediate horizontal edge ...
+        // nb: Smaller Dx's are to the right of larger Dx's ABOVE the horizontal.
+        if ((*bnd)->curr.x == (*horz_bound)->current_edge->top.x && 
+            std::next((*horz_bound)->current_edge) != (*horz_bound)->edges.end() &&
+            (*horz_bound)->current_edge->dx < std::next((*horz_bound)->current_edge)->dx) {
+            break;
+        }
+
+        // note: may be done multiple times
+        if ((*horz_bound)->ring && !is_open) {
+            p1 = add_point_to_ring(horz_bound, (*bnd)->curr);
+            for (auto bnd_itr = active_bounds.begin(); bnd_itr != active_bounds.end(); ++bnd_itr) {
+                if (bnd_itr != horz_bound && (*bnd_itr)->ring && current_edge_is_horizontal<T>(bnd_itr) && 
+                    horizontal_segments_overlap((*horz_bound)->current_edge->bot.x, 
+                                                (*horz_bound)->current_edge->top.x,
+                                                (*bnd_itr)->current_edge->bot.x,
+                                                (*bnd_itr)->current_edge->top.x)) {
+                    point_ptr<T> p2 = get_last_point<T>(bnd_itr);
+                    joins.emplace_back(p2, p1, (*bnd_itr)->current_edge->top);
+                }
+            }
+        }
+
+        // OK, so far we're still in range of the horizontal Edge  but make sure
+        // we're at the last of consec. horizontals when matching with eMaxPair
+        if (is_maxima_edge && bnd == bound_max_pair) {
+            if ((*horz_bound)->ring) {
+                add_local_maximum_point(horz_bound, bound_max_pair, (*horz_bound)->current_edge->top, rings,
+                                        active_bounds);
+            }
+            active_bounds.erase(horz_bound);
+            active_bounds.erase(bound_max_pair);
+            return;
+        }
+
+        intersect_bounds(horz_bound, bnd,
+                        mapbox::geometry::point<T>((*bnd)->curr.x, (*horz_bound)->curr.y), cliptype,
+                        subject_fill_type, clip_fill_type, rings, joins, active_bounds);
+        auto next_bnd = std::next(bnd);
+        active_bounds.splice(horz_bound, active_bounds, bnd);
+        bnd = next_bnd;
+    } // end while (bnd != active_bounds.end())
+
+    if ((*horz_bound)->ring && !p1) {
+        p1 = get_last_point<T>(horz_bound);
+        for (auto bnd_itr = active_bounds.begin(); bnd_itr != active_bounds.end(); ++bnd_itr) {
+            if (bnd_itr != horz_bound && (*bnd_itr)->ring && current_edge_is_horizontal<T>(bnd_itr) && 
+                horizontal_segments_overlap((*horz_bound)->current_edge->bot.x, 
+                                            (*horz_bound)->current_edge->top.x,
+                                            (*bnd_itr)->current_edge->bot.x,
+                                            (*bnd_itr)->current_edge->top.x)) {
+                point_ptr<T> p2 = get_last_point<T>(bnd_itr);
+                joins.emplace_back(p2, p1, (*bnd_itr)->current_edge->top);
+            }
+        }
+    }
+
+    if (std::next((*horz_bound)->current_edge) != (*horz_bound)->edges.end()) {
+        if ((*horz_bound)->ring) {
+            p1 = add_point_to_ring(horz_bound, (*horz_bound)->current_edge->top);
+            if ((*horz_bound)->winding_delta != 0) {
+                auto bnd_prev = active_bound_list_rev_itr<T>(horz_bound);
+                auto bnd_next = std::next(horz_bound);
+                if (bnd_prev != active_bounds.rend() && (*bnd_prev)->ring &&
+                    (*bnd_prev)->curr.x == (*horz_bound)->current_edge->top.x && 
+                    (*bnd_prev)->winding_delta != 0) {
+                    add_point_to_ring(bnd_prev, (*horz_bound)->current_edge->top);
+                }
+                if (bnd_next != active_bounds.end() && (*bnd_next)->ring &&
+                    (*bnd_next)->curr.x == (*horz_bound)->current_edge->top.x && 
+                    (*bnd_next)->winding_delta != 0) {
+                    add_point_to_ring(bnd_next, (*horz_bound)->current_edge->top);
+                }
+            }
+            next_edge_in_bound(horz_bound, scanbeam);
+
+            if ((*horz_bound)->winding_delta == 0) {
+                return;
+            }
+
+            // horz_bound is no longer horizontal here!!! the edge
+            // in the bound was updated above
+            auto bnd_prev = active_bound_list_rev_itr<T>(horz_bound);
+            auto bnd_next = std::next(horz_bound);
+
+            if (bnd_prev != active_bounds.rend() && (*bnd_prev)->curr == (*horz_bound)->current_edge->bot && 
+                (*bnd_prev)->winding_delta != 0 &&
+                (*bnd_prev)->ring && (*bnd_prev)->curr.y > (*bnd_prev)->current_edge->top.y &&
+                 slopes_equal(*((*horz_bound)->current_edge), *((*bnd_prev)->current_edge))) {
+                point_ptr<T> p2 = add_point_to_ring(bnd_prev, (*horz_bound)->current_edge->bot);
+                joins.emplace_back(p1, p2, (*horz_bound)->current_edge->top);
+            } else if (bnd_next != active_bounds.end() && (*bnd_next)->curr == (*horz_bound)->current_edge->bot &&
+                       (*bnd_next)->winding_delta != 0 &&
+                       (*bnd_next)->ring && (*bnd_next)->curr.y > (*bnd_next)->current_edge->top.y &&
+                       slopes_equal(*((*horz_bound)->current_edge), *((*bnd_next)->current_edge))) {
+                point_ptr<T> p2 = add_point_to_ring(bnd_next, (*horz_bound)->current_edge->bot);
+                joins.emplace_back(p1, p2, (*horz_bound)->current_edge->top);
+            }
+        } else {
+            next_edge_in_bound(horz_bound, scanbeam);
+        }
     } else {
-        left = edge->top.x;
-        right = edge->bot.x;
-        dir = right_to_left;
+        if ((*horz_bound)->ring) {
+            add_point_to_ring(horz_bound, (*horz_bound)->current_edge->top);
+        }
+        active_bounds.erase(horz_bound);
     }
 }
 
 template <typename T>
-void process_horizontal(edge_ptr<T> horz_edge,
+void process_horizontal_right_to_left(T scanline_y,
+                            active_bound_list_itr<T> horz_bound,
+                            maxima_list<T>& maxima,
+                            active_bound_list<T>& active_bounds,
+                            join_list<T>& joins,
+                            ring_list<T>& rings,
+                            scanbeam_list<T>& scanbeam,
+                            clip_type cliptype,
+                            fill_type subject_fill_type,
+                            fill_type clip_fill_type) {
+    bool is_open = (*horz_bound)->winding_delta == 0;
+    bool is_maxima_edge = is_maxima(horz_bound, scanline_y);
+    auto bound_max_pair = active_bounds.end();
+    if (is_maxima_edge) {
+        bound_max_pair = get_maxima_pair<T>(horz_bound, active_bounds);
+    }
+
+    auto max_iter = maxima.rbegin();
+    while (max_iter != maxima.rend() && *max_iter > (*horz_bound)->current_edge->bot.x) {
+        ++max_iter;
+    }
+
+    point_ptr<T> p1 = nullptr;
+
+    auto bnd = active_bound_list_rev_itr<T>(horz_bound);
+
+    while (bnd != active_bounds.rend()) {
+        // this code block inserts extra coords into horizontal edges (in output
+        // polygons) wherever maxima touch these horizontal edges. This helps
+        //'simplifying' polygons (ie if the Simplify property is set).
+        while (max_iter != maxima.rend() && *max_iter > (*bnd)->curr.x) {
+            if ((*horz_bound)->ring && !is_open) {
+                add_point_to_ring(horz_bound, mapbox::geometry::point<T>(*max_iter,
+                                                                (*horz_bound)->current_edge->bot.y));
+            }
+            ++max_iter;
+        }
+
+        if ((*bnd)->curr.x < (*horz_bound)->current_edge->top.x) {
+            break;
+        }
+
+        // Also break if we've got to the end of an intermediate horizontal edge ...
+        // nb: Smaller Dx's are to the right of larger Dx's ABOVE the horizontal.
+        if ((*bnd)->curr.x == (*horz_bound)->current_edge->top.x && 
+            std::next((*horz_bound)->current_edge) != (*horz_bound)->edges.end() &&
+            (*horz_bound)->current_edge->dx < std::next((*horz_bound)->current_edge)->dx) {
+            break;
+        }
+
+        // note: may be done multiple times
+        if ((*horz_bound)->ring && !is_open) {
+            p1 = add_point_to_ring(horz_bound, (*bnd)->curr);
+            for (auto bnd_itr = active_bounds.begin(); bnd_itr != active_bounds.end(); ++bnd_itr) {
+                if (bnd_itr != horz_bound && (*bnd_itr)->ring && current_edge_is_horizontal<T>(bnd_itr) && 
+                    horizontal_segments_overlap((*horz_bound)->current_edge->bot.x, 
+                                                (*horz_bound)->current_edge->top.x,
+                                                (*bnd_itr)->current_edge->bot.x,
+                                                (*bnd_itr)->current_edge->top.x)) {
+                    point_ptr<T> p2 = get_last_point<T>(bnd_itr);
+                    joins.emplace_back(p2, p1, (*bnd_itr)->current_edge->top);
+                }
+            }
+        }
+        auto bnd_forward = --(bnd.base());
+
+        // OK, so far we're still in range of the horizontal Edge  but make sure
+        // we're at the last of consec. horizontals when matching with eMaxPair
+        if (is_maxima_edge && bnd_forward == bound_max_pair) {
+            if ((*horz_bound)->ring) {
+                add_local_maximum_point(horz_bound, bound_max_pair, (*horz_bound)->current_edge->top, rings,
+                                        active_bounds);
+            }
+            active_bounds.erase(horz_bound);
+            active_bounds.erase(bound_max_pair);
+            return;
+        }
+
+        intersect_bounds(bnd_forward, horz_bound,
+                        mapbox::geometry::point<T>((*bnd)->curr.x, (*horz_bound)->curr.y), cliptype,
+                        subject_fill_type, clip_fill_type, rings, joins, active_bounds);
+        auto next_bnd = std::next(bnd);
+        active_bounds.splice(bnd_forward, active_bounds, horz_bound);
+        bnd = next_bnd;
+    } // end while (bnd != active_bounds.rend())
+
+    if ((*horz_bound)->ring && !p1) {
+        p1 = get_last_point<T>(horz_bound);
+        for (auto bnd_itr = active_bounds.begin(); bnd_itr != active_bounds.end(); ++bnd_itr) {
+            if (bnd_itr != horz_bound && (*bnd_itr)->ring && current_edge_is_horizontal<T>(bnd_itr) && 
+                horizontal_segments_overlap((*horz_bound)->current_edge->bot.x, 
+                                            (*horz_bound)->current_edge->top.x,
+                                            (*bnd_itr)->current_edge->bot.x,
+                                            (*bnd_itr)->current_edge->top.x)) {
+                point_ptr<T> p2 = get_last_point<T>(bnd_itr);
+                joins.emplace_back(p2, p1, (*bnd_itr)->current_edge->top);
+            }
+        }
+    }
+
+    if (std::next((*horz_bound)->current_edge) != (*horz_bound)->edges.end()) {
+        if ((*horz_bound)->ring) {
+            p1 = add_point_to_ring(horz_bound, (*horz_bound)->current_edge->top);
+            if ((*horz_bound)->winding_delta != 0) {
+                auto bnd_prev = active_bound_list_rev_itr<T>(horz_bound);
+                auto bnd_next = std::next(horz_bound);
+                if (bnd_prev != active_bounds.rend() && (*bnd_prev)->ring &&
+                    (*bnd_prev)->curr.x == (*horz_bound)->current_edge->top.x && 
+                    (*bnd_prev)->winding_delta != 0) {
+                    add_point_to_ring(bnd_prev, (*horz_bound)->current_edge->top);
+                }
+                if (bnd_next != active_bounds.end() && (*bnd_next)->ring &&
+                    (*bnd_next)->curr.x == (*horz_bound)->current_edge->top.x && 
+                    (*bnd_next)->winding_delta != 0) {
+                    add_point_to_ring(bnd_next, (*horz_bound)->current_edge->top);
+                }
+            }
+            next_edge_in_bound(horz_bound, scanbeam);
+
+            if ((*horz_bound)->winding_delta == 0) {
+                return;
+            }
+
+            // horz_bound is no longer horizontal here!!! the edge
+            // in the bound was updated above
+            auto bnd_prev = active_bound_list_rev_itr<T>(horz_bound);
+            auto bnd_next = std::next(horz_bound);
+
+            if (bnd_prev != active_bounds.rend() && (*bnd_prev)->curr == (*horz_bound)->current_edge->bot && 
+                (*bnd_prev)->winding_delta != 0 &&
+                (*bnd_prev)->ring && (*bnd_prev)->curr.y > (*bnd_prev)->current_edge->top.y &&
+                 slopes_equal(*((*horz_bound)->current_edge), *((*bnd_prev)->current_edge))) {
+                point_ptr<T> p2 = add_point_to_ring(bnd_prev, (*horz_bound)->current_edge->bot);
+                joins.emplace_back(p1, p2, (*horz_bound)->current_edge->top);
+            } else if (bnd_next != active_bounds.end() && (*bnd_next)->curr == (*horz_bound)->current_edge->bot &&
+                       (*bnd_next)->winding_delta != 0 &&
+                       (*bnd_next)->ring && (*bnd_next)->curr.y > (*bnd_next)->current_edge->top.y &&
+                       slopes_equal(*((*horz_bound)->current_edge), *((*bnd_next)->current_edge))) {
+                point_ptr<T> p2 = add_point_to_ring(bnd_next, (*horz_bound)->current_edge->bot);
+                joins.emplace_back(p1, p2, (*horz_bound)->current_edge->top);
+            }
+        } else {
+            next_edge_in_bound(horz_bound, scanbeam);
+        }
+    } else {
+        if ((*horz_bound)->ring) {
+            add_point_to_ring(horz_bound, (*horz_bound)->current_edge->top);
+        }
+        active_bounds.erase(horz_bound);
+    }
+}
+
+template <typename T>
+void process_horizontal(T scanline_y,
+                        active_bound_list_itr<T> horz_bound,
                         maxima_list<T>& maxima,
-                        edge_ptr<T>& sorted_edge_list,
-                        edge_ptr<T>& active_edge_list,
+                        active_bound_list<T>& active_bounds,
                         join_list<T>& joins,
-                        join_list<T>& ghost_joins,
                         ring_list<T>& rings,
                         scanbeam_list<T>& scanbeam,
                         clip_type cliptype,
                         fill_type subject_fill_type,
                         fill_type clip_fill_type) {
-    T left;
-    T right;
-    horizontal_direction dir;
-    bool is_open = horz_edge->winding_delta == 0;
-
-    get_horizontal_direction(horz_edge, dir, left, right);
-    edge_ptr<T> last_horizontal = horz_edge;
-    edge_ptr<T> edge_max_pair = nullptr;
-
-    // locate the final consecutive horizontal edge starting from this one;
-    // multiple horizontals in a row will be linked
-    while (last_horizontal->next_in_LML && is_horizontal(*(last_horizontal->next_in_LML))) {
-        last_horizontal = last_horizontal->next_in_LML;
-    }
-
-    if (!last_horizontal->next_in_LML)
-        edge_max_pair = get_maxima_pair(last_horizontal);
-
-    typename maxima_list<T>::const_iterator max_iter;
-    typename maxima_list<T>::const_reverse_iterator max_reverse_iter;
-
-    if (!maxima.empty()) {
-        // get the first maxima in range (X) ...
-        if (dir == left_to_right) {
-            max_iter = maxima.begin();
-            while (max_iter != maxima.end() && *max_iter <= horz_edge->bot.x) {
-                ++max_iter;
-            }
-            if (max_iter != maxima.end() && *max_iter >= last_horizontal->top.x) {
-                max_iter = maxima.end();
-            }
-        } else {
-            max_reverse_iter = maxima.rbegin();
-            while (max_reverse_iter != maxima.rend() && *max_reverse_iter > horz_edge->bot.x) {
-                ++max_reverse_iter;
-            }
-            if (max_reverse_iter != maxima.rend() && *max_reverse_iter <= last_horizontal->top.x) {
-                max_reverse_iter = maxima.rend();
-            }
-        }
-    }
-
-    point_ptr<T> p1 = nullptr;
-
-    // loop through consec. horizontal edges
-    for (;;) {
-        bool is_last_horizontal = (horz_edge == last_horizontal);
-        edge_ptr<T> e = get_next_in_AEL(horz_edge, dir);
-
-        while (e) {
-            // this code block inserts extra coords into horizontal edges (in output
-            // polygons) wherever maxima touch these horizontal edges. This helps
-            //'simplifying' polygons (ie if the Simplify property is set).
-            if (!maxima.empty()) {
-                if (dir == left_to_right) {
-                    while (max_iter != maxima.end() && *max_iter < e->curr.x) {
-                        if (horz_edge->index >= 0 && !is_open) {
-                            add_point(horz_edge,
-                                      mapbox::geometry::point<T>(*max_iter, horz_edge->bot.y),
-                                      rings);
-                        }
-                        ++max_iter;
-                    }
-                } else {
-                    while (max_reverse_iter != maxima.rend() && *max_reverse_iter > e->curr.x) {
-                        if (horz_edge->index >= 0 && !is_open) {
-                            add_point(horz_edge, mapbox::geometry::point<T>(*max_reverse_iter,
-                                                                            horz_edge->bot.y),
-                                      rings);
-                        }
-                        ++max_reverse_iter;
-                    }
-                }
-            }
-
-            if ((dir == left_to_right && e->curr.x > right) ||
-                (dir == right_to_left && e->curr.x < left)) {
-                break;
-            }
-
-            // Also break if we've got to the end of an intermediate horizontal edge ...
-            // nb: Smaller Dx's are to the right of larger Dx's ABOVE the horizontal.
-            if (e->curr.x == horz_edge->top.x && horz_edge->next_in_LML &&
-                e->dx < horz_edge->next_in_LML->dx) {
-                break;
-            }
-
-            // note: may be done multiple times
-            if (horz_edge->index >= 0 && !is_open) {
-                p1 = add_point(horz_edge, e->curr, rings);
-                edge_ptr<T> e_next_horizontal = sorted_edge_list;
-
-                while (e_next_horizontal) {
-                    if (e_next_horizontal->index >= 0 &&
-                        horizontal_segments_overlap(horz_edge->bot.x, horz_edge->top.x,
-                                                    e_next_horizontal->bot.x,
-                                                    e_next_horizontal->top.x)) {
-                        point_ptr<T> p2 = get_last_point(e_next_horizontal, rings);
-                        joins.emplace_back(p2, p1, e_next_horizontal->top);
-                    }
-                    e_next_horizontal = e_next_horizontal->next_in_SEL;
-                }
-                ghost_joins.emplace_back(p1, nullptr, horz_edge->bot);
-            }
-
-            // OK, so far we're still in range of the horizontal Edge  but make sure
-            // we're at the last of consec. horizontals when matching with eMaxPair
-            if (e == edge_max_pair && is_last_horizontal) {
-                if (horz_edge->index >= 0) {
-                    add_local_maximum_point(horz_edge, edge_max_pair, horz_edge->top, rings,
-                                            active_edge_list);
-                }
-                delete_from_AEL(horz_edge, active_edge_list);
-                delete_from_AEL(edge_max_pair, active_edge_list);
-                return;
-            }
-
-            if (dir == left_to_right) {
-                intersect_edges(horz_edge, e,
-                                mapbox::geometry::point<T>(e->curr.x, horz_edge->curr.y), cliptype,
-                                subject_fill_type, clip_fill_type, rings, joins, active_edge_list);
-            } else {
-                intersect_edges(e, horz_edge,
-                                mapbox::geometry::point<T>(e->curr.x, horz_edge->curr.y), cliptype,
-                                subject_fill_type, clip_fill_type, rings, joins, active_edge_list);
-            }
-
-            edge_ptr<T> e_next = get_next_in_AEL(e, dir);
-            swap_positions_in_AEL(horz_edge, e, active_edge_list);
-            e = e_next;
-        } // end while (e)
-
-        // Break out of loop if HorzEdge.NextInLML is not also horizontal ...
-        if (!horz_edge->next_in_LML || !is_horizontal(*(horz_edge->next_in_LML))) {
-            break;
-        }
-
-        update_edge_into_AEL(horz_edge, active_edge_list, scanbeam);
-        if (horz_edge->index >= 0) {
-            add_point(horz_edge, horz_edge->bot, rings);
-        }
-        get_horizontal_direction(horz_edge, dir, left, right);
-    } // end for (;;)
-
-    if (horz_edge->index >= 0 && !p1) {
-        p1 = get_last_point(horz_edge, rings);
-        edge_ptr<T> e_next_horizontal = sorted_edge_list;
-        while (e_next_horizontal) {
-            if (e_next_horizontal->index >= 0 &&
-                horizontal_segments_overlap(horz_edge->bot.x, horz_edge->top.x,
-                                            e_next_horizontal->bot.x, e_next_horizontal->top.x)) {
-                point_ptr<T> p2 = get_last_point(e_next_horizontal, rings);
-                joins.emplace_back(p2, p1, e_next_horizontal->top);
-            }
-            e_next_horizontal = e_next_horizontal->next_in_SEL;
-        }
-        ghost_joins.emplace_back(p1, nullptr, horz_edge->top);
-    }
-
-    if (horz_edge->next_in_LML) {
-        if (horz_edge->index >= 0) {
-            p1 = add_point(horz_edge, horz_edge->top, rings);
-            edge_ptr<T> e_prev = horz_edge->prev_in_AEL;
-            edge_ptr<T> e_next = horz_edge->next_in_AEL;
-
-            if ((horz_edge->winding_delta != 0) && e_prev && (e_prev->index >= 0) &&
-                (e_prev->curr.x == horz_edge->top.x) && (e_prev->winding_delta != 0)) {
-                add_point(e_prev, horz_edge->top, rings);
-            }
-            if ((horz_edge->winding_delta != 0) && e_next && (e_next->index >= 0) &&
-                (e_next->curr.x == horz_edge->top.x) && (e_next->winding_delta != 0)) {
-                add_point(e_next, horz_edge->top, rings);
-            }
-            update_edge_into_AEL(horz_edge, active_edge_list, scanbeam);
-
-            if (horz_edge->winding_delta == 0) {
-                return;
-            }
-
-            // nb: HorzEdge is no longer horizontal here
-            e_prev = horz_edge->prev_in_AEL;
-            e_next = horz_edge->next_in_AEL;
-
-            if (e_prev && e_prev->curr.x == horz_edge->bot.x &&
-                e_prev->curr.y == horz_edge->bot.y && e_prev->winding_delta != 0 &&
-                (e_prev->index >= 0 && e_prev->curr.y > e_prev->top.y &&
-                 slopes_equal(*horz_edge, *e_prev))) {
-                point_ptr<T> p2 = add_point(e_prev, horz_edge->bot, rings);
-                joins.emplace_back(p1, p2, horz_edge->top);
-            } else if (e_next && e_next->curr.x == horz_edge->bot.x &&
-                       e_next->curr.y == horz_edge->bot.y && e_next->winding_delta != 0 &&
-                       e_next->index >= 0 && e_next->curr.y > e_next->top.y &&
-                       slopes_equal(*horz_edge, *e_next)) {
-                point_ptr<T> p2 = add_point(e_next, horz_edge->bot, rings);
-                joins.emplace_back(p1, p2, horz_edge->top);
-            }
-        } else {
-            update_edge_into_AEL(horz_edge, active_edge_list, scanbeam);
-        }
+    if ((*horz_bound)->current_edge->bot.x < (*horz_bound)->current_edge->top.x) {
+        process_horizontal_left_to_right(scanline_y,
+                            horz_bound,
+                            maxima,
+                            active_bounds,
+                            joins,
+                            rings,
+                            scanbeam,
+                            cliptype,
+                            subject_fill_type,
+                            clip_fill_type);
     } else {
-        if (horz_edge->index >= 0) {
-            add_point(horz_edge, horz_edge->top, rings);
-        }
-        delete_from_AEL(horz_edge, active_edge_list);
+        process_horizontal_right_to_left(scanline_y,
+                            horz_bound,
+                            maxima,
+                            active_bounds,
+                            joins,
+                            rings,
+                            scanbeam,
+                            cliptype,
+                            subject_fill_type,
+                            clip_fill_type);
     }
+}
+
+template <typename T>
+void process_horizontals(T scanline_y,
+                         maxima_list<T>& maxima,
+                         active_bound_list<T>& active_bounds,
+                         join_list<T>& joins,
+                         ring_list<T>& rings,
+                         scanbeam_list<T>& scanbeam,
+                         clip_type cliptype,
+                         fill_type subject_fill_type,
+                         fill_type clip_fill_type) {
+    maxima.sort();
+    for (auto bnd_itr = active_bounds.begin(); bnd_itr != active_bounds.end(); ++bnd_itr) {
+        if (current_edge_is_horizontal<T>(bnd_itr)) {
+            process_horizontal(scanline_y, bnd_itr, maxima, active_bounds, joins,
+                               rings, scanbeam, cliptype, subject_fill_type,
+                               clip_fill_type);
+        }
+    }
+    maxima.clear();
 }
 }
 }
